+++ scripts/build_renewable_profiles.py
@@ -24,8 +24,9 @@ Relevant settings
         {technology}:
             cutout:
             copernicus:
-            grid_codes:
-            distance:
+                grid_codes:
+                distance:
+                distance_grid_codes:
             natura:
             max_depth:
             max_shore_distance:
@@ -45,7 +46,7 @@ Relevant settings
 Inputs
 ------
 
-- ``data/raw/copernicus/PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif``: `Copernicus Land Service <https://land.copernicus.eu/global/products/lc>`_ inventory on `32 classes <http://www.fao.org/3/x0596e/X0596e01i.htm#P1473_136884>`_ of land use (e.g. forests, arable land, industrial, urban areas) based on UN-FAO calssification.
+- ``data/raw/copernicus/PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif``: `Copernicus Land Service <https://land.copernicus.eu/global/products/lc>`_ inventory on 23 land use classes (e.g. forests, arable land, industrial, urban areas) based on UN-FAO classification. See `Table 4 in the PUM <https://land.copernicus.eu/global/sites/cgls.vito.be/files/products/CGLOPS1_PUM_LC100m-V3_I3.4.pdf>`_ for a list of all classes.
 
     .. image:: ../img/copernicus.png
         :scale: 33 %
@@ -287,20 +288,22 @@ if __name__ == "__main__":
         if "natura" in config and config["natura"]:
             excluder.add_raster(paths.natura, nodata=0, allow_no_overlap=True)
 
-        copernicus = config.get("copernicus", {})
-        if "grid_codes" in copernicus:
-            codes = copernicus["grid_codes"]
-            excluder.add_raster(paths.copernicus,
-                                codes=codes,
-                                invert=True,
-                                crs=3035)
-        if copernicus.get("distance", 0.0) > 0.0:
-            codes = copernicus["distance_grid_codes"]
-            buffer = copernicus["distance"]
-            excluder.add_raster(paths.copernicus,
-                                codes=codes,
-                                buffer=buffer,
-                                crs=3035)
+        if "copernicus" in config and config["copernicus"]:
+            copernicus = config["copernicus"]
+            excluder.add_raster(
+                paths.copernicus,
+                codes=copernicus["grid_codes"],
+                invert=True,
+                crs="EPSG:4236",
+            )
+            if "distance" in copernicus and config["copernicus"][
+                    "distance"] > 0:
+                excluder.add_raster(
+                    paths.copernicus,
+                    codes=copernicus["distance_grid_codes"],
+                    buffer=copernicus["distance"],
+                    crs="EPSG:4236",
+                )
 
         if "max_depth" in config:
             # lambda not supported for atlite + multiprocessing