+++ doc/release_notes.rst
@@ -25,6 +25,8 @@ Upcoming Release
 
 * Fix random state for kmean clustering, `adopted from PR <https://github.com/PyPSA/pypsa-eur/pull/313>`
 
+* Implement area exclusions based on land type using the Copernicus Land Cover (PR #272).
+
 
 PyPSA-Africa 0.0.1 (24th December 2021)
 =====================================